Divide to_move_test_icon_service_engine test file into integrate test and unittest files
