+++ tests/unittest/test_icon_service_engine.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2018 ICON Foundation
+# Copyright 2020 ICON Foundation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,31 +14,43 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""IconScoreEngine testcase
-"""
-from unittest import mock
+"""IconServiceEngine testCase"""
+from unittest.mock import Mock
 
 import pytest
 
-from iconservice.icon_constant import ConfigKey
+from iconservice.icon_constant import IconServiceFlag, ConfigKey
 from iconservice.icon_service_engine import IconServiceEngine
 
 
-def mock_config_get(_, key):
-    return key
+@pytest.fixture
+def engine():
+    engine = IconServiceEngine()
+    return engine
 
 
-# FIXME TODO
-class TestIconServiceEngine:
+@pytest.mark.parametrize("fee", [True, False])
+@pytest.mark.parametrize("audit", [True, False])
+@pytest.mark.parametrize("validator", [True, False])
+def test_make_flag(engine, fee, audit, validator):
+    table = {
+        ConfigKey.SERVICE_FEE: fee,
+        ConfigKey.SERVICE_AUDIT: audit,
+        ConfigKey.SERVICE_SCORE_PACKAGE_VALIDATOR: validator
+    }
+    fee_flag_value = IconServiceFlag.FEE.value if fee else 0
+    audit_flag_value = IconServiceFlag.AUDIT.value if audit else 0
+    validator_flag_value = IconServiceFlag.SCORE_PACKAGE_VALIDATOR.value if validator else 0
+    flag = engine._make_service_flag(table)
+    assert flag == (fee_flag_value | audit_flag_value | validator_flag_value)
 
-    def mock_conf(self, mocker):
-        conf = mock.Mock()
-        conf.attach_mock(mock.Mock(side_effect=lambda key: key), "__getitem__")
-        return conf
 
-    # @pytest.mark.skip("TODO")
-    def test_open(self, mocker, mock_conf):
-        mocker.patch_object(IconServiceEngine, "_make_service_flag", autospec=True)
-
-        engine = IconServiceEngine()
-        engine.open(mock_conf)
+@pytest.mark.parametrize("method", [
+    "icx_getBalance", "icx_getTotalSupply", "icx_call", "icx_sendTransaction",
+    "debug_estimateStep", "icx_getScoreApi", "ise_getStatus"])
+def test_call_method(engine, method):
+    call_method = engine._handlers[method] = Mock()
+    ctx = Mock()
+    params = Mock(spec=dict)
+    engine._call(ctx, method, params)
+    call_method.assert_called_with(ctx, params)
